iss-907: implemented fallback and cache
