+++ hystrix-contrib/hystrix-javanica/src/ajcTest/java/com/netflix/hystrix/contrib/javanica/test/aspectj/collapser/CollapserTest.java
@@ -1,0 +1,19 @@
+package com.netflix.hystrix.contrib.javanica.test.aspectj.collapser;
+
+import com.netflix.hystrix.contrib.javanica.test.common.collapser.BasicCollapserTest;
+import org.junit.BeforeClass;
+
+/**
+ * Created by dmgcodevil
+ */
+public class CollapserTest extends BasicCollapserTest {
+    @BeforeClass
+    public static void setUpEnv() {
+        System.setProperty("weavingMode", "compile");
+    }
+
+    @Override
+    protected UserService createUserService() {
+        return new UserService();
+    }
+}
