+++ hystrix-contrib/hystrix-javanica/src/main/java/com/netflix/hystrix/contrib/javanica/aop/aspectj/HystrixCacheAspect.java
@@ -31,6 +31,9 @@ import org.aspectj.lang.annotation.Pointcut;
 import java.lang.reflect.Method;
 
 import static com.netflix.hystrix.contrib.javanica.utils.AopUtils.getMethodFromTarget;
+import static com.netflix.hystrix.contrib.javanica.utils.EnvUtils.getWeavingMode;
+import static com.netflix.hystrix.contrib.javanica.utils.EnvUtils.isCompileWeaving;
+import static com.netflix.hystrix.contrib.javanica.utils.ajc.AjcUtils.getOriginalMethod;
 
 /**
  * AspectJ aspect to process methods which annotated with annotations from
@@ -54,10 +57,14 @@ public class HystrixCacheAspect {
         MetaHolder metaHolder = MetaHolder.builder()
                 .args(args).method(method).obj(obj)
                 .executionType(ExecutionType.SYNCHRONOUS)
+                .weavingMode(getWeavingMode())
+                .ajcMethod(isCompileWeaving() ? getOriginalMethod(obj.getClass(), method) : null)
                 .build();
         CacheInvocationContext<CacheRemove> context = CacheInvocationContextFactory
                 .createCacheRemoveInvocationContext(metaHolder);
         HystrixRequestCacheManager.getInstance().clearCache(context);
         return joinPoint.proceed();
     }
+
+
 }
