+++ hystrix-contrib/hystrix-javanica/src/main/java/com/netflix/hystrix/contrib/javanica/command/LazyCommandExecutionAction.java
@@ -73,6 +73,8 @@ public class LazyCommandExecutionAction extends CommandAction {
         return MetaHolder.builder()
                 .obj(metaHolder.getObj())
                 .method(metaHolder.getMethod())
+                .ajcMethod(metaHolder.getAjcMethod())
+                .weavingMode(metaHolder.getWeavingMode())
                 .executionType(executionType)
                 .args(metaHolder.getArgs())
                 .defaultCollapserKey(metaHolder.getDefaultCollapserKey())
@@ -87,6 +89,8 @@ public class LazyCommandExecutionAction extends CommandAction {
                 .obj(metaHolder.getObj())
                 .method(metaHolder.getMethod())
                 .executionType(executionType)
+                .ajcMethod(metaHolder.getAjcMethod())
+                .weavingMode(metaHolder.getWeavingMode())
                 .args(args)
                 .defaultCollapserKey(metaHolder.getDefaultCollapserKey())
                 .defaultCommandKey(metaHolder.getDefaultCommandKey())
